feat: implement AI vulnerability assessment page and reusable API testing tool component
