+++ frontend/src/app/assess/[projectId]/vulnerability-assessment/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import ApiTestingTool from "@/components/features/ai-testing/ApiTestingTool";
+
+export default function VulnerabilityAssessmentPage() {
+  return <ApiTestingTool mode="vulnerability" />;
+}
